[#1290] helm charts: introducing pullPolicy
In order to let users customize the behavior of their deployment, it
make sense to introduce a pullPolicy in our helm charts. With using a
default value of null in the base chart we won't impose any particular
behavior, but instead opening a nice customization path for everyone who
needs it.